+++ prep/src/tools/mod.rs
@@ -1,0 +1,5 @@
+// Copyright 2026 the Prep Authors
+// SPDX-License-Identifier: Apache-2.0 OR MIT
+
+pub mod cargo;
+pub mod rustup;