+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 * `init` command to set up custom per-project Prep configuration. ([#23] by [@xStrom])
 * `copyright` command to easily verify that all Rust source files have correct copyright headers. ([#22] by [@xStrom])
 * `--crates <main|aux|all>` option to the `clippy` command. ([#18] by [@xStrom])
+* Automatic Cargo installation via `rustup`. ([#24] by [@xStrom])
 * Ability to run from within a sub-directory of a workspace as opposed to just from the workspace root. ([#23] by [@xStrom])
 
 ### Changed
@@ -29,6 +30,7 @@
 [#18]: https://github.com/Nevermore/prep/pull/18
 [#22]: https://github.com/Nevermore/prep/pull/22
 [#23]: https://github.com/Nevermore/prep/pull/23
+[#24]: https://github.com/Nevermore/prep/pull/24
 
 [Unreleased]: https://github.com/Nevermore/prep/compare/v0.1.0...HEAD
 [0.1.0]: https://github.com/Nevermore/prep/compare/v0.0.0...v0.1.0